resolved uint cmp warnings

+++ src/mysf.cpp
@@ -16,8 +16,8 @@ bool rgbToGray(sf::Image& image, sf::IntRect dimension){
         dimension.width = image.getSize().x;
         dimension.left = dimension.top = 0;
     }
-    for(unsigned int i = dimension.top ;i <  dimension.top + dimension.height; i++)
-        for(unsigned int j = dimension.left ; j <  dimension.left + dimension.width ; j++){
+    for(unsigned int i = dimension.top ;i <  (uint32_t)dimension.top + dimension.height; i++)
+        for(unsigned int j = dimension.left ; j <  (uint32_t)dimension.left + dimension.width ; j++){
             if(j >= image.getSize().x or i >= image.getSize().y)
                 return false;
             else{
@@ -38,8 +38,8 @@ bool grayToMono(sf::Image& image, unsigned int threshold,sf::IntRect dimension){
         dimension.width = image.getSize().x;
         dimension.left = dimension.top = 0;
     }
-    for(unsigned int i = dimension.top ;i < dimension.top +  dimension.height; i++)
-        for(unsigned int j = dimension.left ; j < dimension.left + dimension.width ; j++){
+    for(unsigned int i = (uint32_t)dimension.top ;i < (uint32_t)dimension.top +  dimension.height; i++)
+        for(unsigned int j = dimension.left ; j < (uint32_t)dimension.left + dimension.width ; j++){
             if(image.getPixel(j,i).r > threshold)
                 image.setPixel(j,i,sf::Color(255,255,255,255));
             else
@@ -56,8 +56,8 @@ double mean(sf::Image& image, sf::IntRect dimension){
     } 
     unsigned int height = dimension.height + dimension.top , width = dimension.width + dimension.left;
     Eigen::MatrixXd image_matrix = Eigen::MatrixXd::Zero(height,width);
-    for(unsigned int i = dimension.top ;i < dimension.top +  dimension.height; i++)
-        for(unsigned int j = dimension.left ; j < dimension.left + dimension.width ; j++)
+    for(unsigned int i = dimension.top ;i < (uint32_t)dimension.top +  dimension.height; i++)
+        for(unsigned int j = dimension.left ; j < (uint32_t)dimension.left + dimension.width ; j++)
             image_matrix(i,j) = image.getPixel(j,i).r;
     return image_matrix.mean();
 }
@@ -70,8 +70,8 @@ double variance(sf::Image& image,sf::IntRect dimension){
     } 
     unsigned int height = dimension.height + dimension.top , width = dimension.width + dimension.left;
     Eigen::MatrixXd image_matrix = Eigen::MatrixXd::Zero(height,width);
-    for(unsigned int i = dimension.top ;i < dimension.top +  dimension.height; i++)
-        for(unsigned int j = dimension.left ; j < dimension.left + dimension.width ; j++)
+    for(unsigned int i = dimension.top ;i < (uint32_t)dimension.top +  dimension.height; i++)
+        for(unsigned int j = dimension.left ; j < (uint32_t)dimension.left + dimension.width ; j++)
             image_matrix(i,j) = image.getPixel(j,i).r;
     Eigen::MatrixXd transformed_matrix = (image_matrix - 
                 Eigen::MatrixXd::Constant(dimension.height,dimension.width,image_matrix.mean()));
